requirements.txt
fastapi
uvicorn
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic-settings
python-dotenv
google-cloud-aiplatform
python-multipart
sqlalchemy
python-decouple